Déplacement de l'algo de chargement de branche modèle + localisation
